use getEnumerationValue on slatAngleType

+++ tst/EnergyPlus/unit/WindowEquivalentLayer.unit.cc
@@ -189,7 +189,7 @@ TEST_F(EnergyPlusFixture, WindowEquivalentLayer_GetInput)
     }
     EXPECT_EQ(1, state->dataHeatBal->TotBlindsEQL);
     EXPECT_TRUE(compare_enums(state->dataMaterial->Material(VBMatNum)->Group, Material::MaterialGroup::BlindEquivalentLayer));
-    EXPECT_EQ(state->dataMaterial->Material(VBMatNum)->SlatAngleType, state->dataWindowEquivalentLayer->lscVBNOBM);
+    EXPECT_EQ(static_cast<int>(state->dataMaterial->Material(VBMatNum)->SlatAngleType), state->dataWindowEquivalentLayer->lscVBNOBM);
 
     int ConstrNum = 1;
     int EQLNum = 0;
@@ -552,7 +552,7 @@ TEST_F(EnergyPlusFixture, WindowEquivalentLayer_VBMaximizeBeamSolar)
     // get equivalent layer window optical properties
     CalcEQLOpticalProperty(*state, SurfNum, DataWindowEquivalentLayer::SolarArrays::BEAM, AbsSolBeam);
     // check that the slat angle control type is set to MaximizeSolar
-    EXPECT_EQ(state->dataMaterial->Material(VBMatNum)->SlatAngleType, state->dataWindowEquivalentLayer->lscVBPROF);
+    EXPECT_EQ(static_cast<int>(state->dataMaterial->Material(VBMatNum)->SlatAngleType), state->dataWindowEquivalentLayer->lscVBPROF);
     // check the slat angle
     EXPECT_NEAR(-71.0772, state->dataSurface->SurfWinSlatAngThisTSDeg(SurfNum), 0.0001);
     // check that for MaximizeSolar slat angle control, the slat angle = -ve vertical profile angle
@@ -913,7 +913,7 @@ TEST_F(EnergyPlusFixture, WindowEquivalentLayer_VBBlockBeamSolar)
     // calc window optical property
     CalcEQLOpticalProperty(*state, SurfNum, DataWindowEquivalentLayer::SolarArrays::BEAM, AbsSolBeam);
     // check VB slat angle for BlockBeamSolar slat angle control
-    EXPECT_EQ(state->dataMaterial->Material(VBMatNum)->SlatAngleType, state->dataWindowEquivalentLayer->lscVBNOBM);
+    EXPECT_EQ(static_cast<int>(state->dataMaterial->Material(VBMatNum)->SlatAngleType), state->dataWindowEquivalentLayer->lscVBNOBM);
     // check the VB slat angle
     EXPECT_NEAR(18.9228, state->dataSurface->SurfWinSlatAngThisTSDeg(SurfNum), 0.0001);
     // check that for BlockBeamSolar slat angle control, the slat angle = 90 - ProfAngVer
@@ -1997,7 +1997,7 @@ TEST_F(EnergyPlusFixture, WindowEquivalentLayer_VBEffectiveEmissivityTest)
         }
     }
     // check VB slat angle control for FixedSlatAngle
-    EXPECT_EQ(state->dataMaterial->Material(VBMatNum)->SlatAngleType, state->dataWindowEquivalentLayer->lscNONE);
+    EXPECT_EQ(static_cast<int>(state->dataMaterial->Material(VBMatNum)->SlatAngleType), state->dataWindowEquivalentLayer->lscNONE);
 
     EQLNum = state->dataConstruction->Construct(ConstrNum).EQLConsPtr;
     // check number of solid layers
